+++ eleanor/output/csv.py
@@ -484,6 +484,10 @@ class CsvSink(AbstractOutputSink):
         return False
 
     @override
+    def supports_background_commit(self) -> bool:
+        return True
+
+    @override
     def supports_progress(self) -> bool:
         return True