+++ eleanor/output/memory.py
@@ -100,6 +100,10 @@ class MemorySink(AbstractOutputSink):
         return self.settings.support_worker_commit
 
     @override
+    def supports_background_commit(self) -> bool:
+        return True
+
+    @override
     def supports_progress(self) -> bool:
         return True